split out fileio utils
